Perform type checking on user library usage

+++ tests/scripts/paper_eval/cache_sim/cache_sim.mm
@@ -1,5 +1,5 @@
 use cache_sim;
 
 wasm:opcode:i32_load:before {
-    cache_sim.load(effective_addr);
+    cache_sim.load(effective_addr as i32);
 }
